+++ OS/01/conversations/user.json
@@ -1,1 +1,1 @@
-[{"role": "user", "type": "message", "content": "Hi.\n"}]
+[{"role": "user", "type": "message", "content": "Hi, can you hear me?\n"}]
