+++ OS/01/utils/kernel.py
@@ -1,13 +1,6 @@
-"""
-Watches the kernel. When it sees something that passes a filter,
-it sends POST request with that to /computer.
-"""
-
+import asyncio
 import subprocess
-import time
-import requests
 import platform
-import os
 
 def get_kernel_messages():
     """
@@ -43,13 +36,28 @@ def custom_filter(message):
 last_messages = ""
 
 def check_filtered_kernel():
+    messages = get_kernel_messages()
+    messages.replace(last_messages, "")
+    messages = messages.split("\n")
+    
+    filtered_messages = []
+    for message in messages:
+        if custom_filter(message):
+            filtered_messages.append(message)
+    
+    return "\n".join(filtered_messages)
+
+async def put_kernel_messages_into_queue(queue):
     while True:
-        messages = get_kernel_messages()
-        messages.replace(last_messages, "")
-        messages = messages.split("\n")
+        text = check_filtered_kernel()
+        if text:
+            if isinstance(queue, asyncio.Queue):
+                await queue.put({"role": "computer", "type": "console", "start": True})
+                await queue.put({"role": "computer", "type": "console", "format": "output", "content": text})
+                await queue.put({"role": "computer", "type": "console", "end": True})
+            else:
+                queue.put({"role": "computer", "type": "console", "start": True})
+                queue.put({"role": "computer", "type": "console", "format": "output", "content": text})
+                queue.put({"role": "computer", "type": "console", "end": True})
         
-        filtered_messages = []
-        for message in messages:
-            if custom_filter(message):
-                filtered_messages.append(message)
-        return "\n".join(filtered_messages)
+        await asyncio.sleep(5)
